refactor(frontend): move create trial modal into separate folder
