+++ frontend/src/components/dashboard/trial/create-trial-modal/CreateTrialModal.module.scss
